[MJS-115] Create logout functionality
The thing we applied is that when a user logs out, we take the token we obtained and push that token in a blackedlisted token list in our DB so that in future if someone tries to login with that token, the person cannot login.

Furthermore we also store the expiration date of each token so that when a particular token expires we can remove that from the database (database cleaning) as JWT will automatically invalidate that particular token when it expires.